Refactor progress screen structure
